Update Todolist.jsx
Added the ability to mark tasks as completed, upon clicking on the task text. When marking a task as completed, it gets a green checkmark next to it. Also, fixed a problem with moving tasks up and down, and their time being switched with the task. Now, when you move tasks, only the tasks text gets moved, the time stays in place.